django/requirements/base.txt
Django==2.2
django-behaviors==0.4.1
django-environ==0.4.5
django-filter==2.1.0
djangorestframework==3.9.2

psycopg2==2.8.1

requests==2.21.0
